refactor(host-application): split runtime orchestrator into focused modules
Extract runtime prerequisite resolution, startup telemetry/process boot, and runtime registry lifecycle into dedicated modules under app_service/runtime_orchestrator.

Keep AppService entrypoints thin while preserving existing runtime behavior and error propagation paths.
